Use correct encoding for line terminator when reading CSV
Using `lineterminator.encode()` will almost-always produce a single byte
encoding, which will break things if you have a 2-byte encoding like
UTF-16-LE. It magically works with UTF-16-BE, because in that case the
additional byte is before the encoded line terminator.

However, one has to be careful when encoding the line terminator, as it
is intended to be in the middle of the file. But encodings that have a
BOM will unintentionally insert the BOM to the encoded line terminator,
causing line splitting to not work at all.