Reinstate try/except block in _bind_property
The change in #11035 to adapt _bind_property to Python 3.11.9 moved
attribute access on the pandas class object around and therefore
dropped exception handling. While it is not the case that we currently
wrap attributes that don't exist, let's avoid an accidental breakage
in the future by reinstating the old exception handling behaviour.