Added key_max_length to assert_eq
In https://github.com/dask/dask-ml/pull/1008 we have a key with a long, but IMO reasonable, name:

```
'frompandas-getitem-operation-assign-values-from-dask-array-getitem-from-dask-array-b02111402bccb70b6fde7347c3d8368f'
```

that's the result of `enc.inverse_transform(enc.transform(df).values))`
where both transform and inverse_transform do some dask.dataframe operations.

If people agree that that name is long but reasonable, then it'd be nice to have the check
in assert_eq be configurable.