Workaround failing shuffle test
This was failing thanks to some poor interaction with meta inference
and what the test was testing. The meta inference for
`Series[str].map(dict)` was (probably correctly?) inferring `object`
dtype. On this PR I've proposed using actual objects for `object` dtype,
which we can't do ordering operations on, which caused the `set_index` to
fail for unreleated reasons.