TST: Fix test_enforce_columns on Python 3.14
Formerly, the `ValueError` would be raised by `DataFrame.__len__` from
the tuple unpacking into the function arguments. In Python 3.14, as of
https://github.com/python/cpython/commit/5a23994a3dbee43a0b08f5920032f60f38b63071
the `__len__` method is not called and the `ValueError` is not raised.

In any case, I don't believe the point of the test is to check `__len__`
due to tuple unpacking, but just general column consistency checks. It
is thus sufficient to call `dask.compute` on the `DataFrame` as a whole,
which will run a column consistency check in all cases.